Implements an adapter for ext-mongodb
This patch is an alternative to #132, proposing a parallel adapter for
ext-mongodb, instead of repurposing the existing ext-mongo adapter to
only work with ext-mongodb.

It copies the various `MongoDb*` classes to `ExtMongoDb*` variants, and
applies the changes as suggested in #132. Additionally, it sets up the
test harness to test ext-mongodb support in PHP 7.1 and 7.2, while
testing ext-mongo support in PHP 5.6 and 7.0. When it tests ext-mongodb,
it also adds a dependency on mongodb/mongodb so that the client library
is available.